Delete Max%/2_Intermediate/pointless_machines_S.tas
just like the name suggest, this file is kinda pointless. The standard any% TAS inputs pick up the silver anyways, so in the context of the full intermediate lobby run, the level can just be read from any% files directly

If this ever changes, or if a deathwarp is added later, then the file does have a purposes and can be added back in relatively easily.