-453f 74.tas (1:42.714) from amberlovesarchie
https://discord.com/channels/1097269700272083115/1097270459826978876/1259766475862970439
-453f 74 (1:50.415(6495) -> 1:42.714(6042))
add new any% stuff and redo half the berries - much faster route for 1st berry, redo 2nd berry and cut the deathwarp.  still a lot to look at (i couldn't be bothered to look at the 3rd and 4th berry rooms) but improved most of the transitional movement pretty significantly.  also read commands need to be added for a lot of this TAS i'll get around to it once i touch up the rest of this next push