Intermediate Lobby All Cats All Challenges (ACAC) Draft (33:53.387)
In terms of the levels, most of them are the same as any%, except for, in order:

Supernautica
Honeyzip Inc.
In Filtration
Square the Circle
Frosted Fragments

In terms of the lobby, the routing is exactly the same as max% (not incl level re-entries) except in 2 spots:
at start, instead of going right to fifth dimension, the TAS goes and enters the gym. the gym is completed by returning to the lobby via the option in the hub to do so (could also RTL via menu) but this is more "concrete"
up until after the final level (low-G botany), where the TAS goes and pets the cat thats to the right of the rainbow podium

I draftedHoneyzip Inc., In Filtration, and Square the Circle
I also did the max% int routing and drafted the lobby cat file
Mirkwood drafted Supernautica and Frosted Fragments, so huge thanks to them!