Beginner Lobby All Cats All Challenges (ACAC) Draft (30:38.465)
In terms of levels, the ones that are different from any% are:
Troposphere
A Gift from the Stars
Seeing is Believing
Dropzle
Strawberry Orchard
Paint
Blueberry Bay

In terms of the lobby, pretty much everything has been rerouted. Mainly because of ONE THING: The Crab Cat, which requires bringing a crab (reskinned theo crystal) from the top-right of the lobby near Over the City alcove ALL THE WAY to the bottom-left near the Midnight Spire alcove.
this route also needs to keep in mind a decent place to pet the Bamboo cat, and, as usual, entering the gym at the beginning (which thankfully connects quite nicely to loopy lagoon which is the first any% level played)
I tested a few routes with the knowledge that one lobby connection has to be Over the City to Midnight Spire, and I think I did pretty good all things considered

I drafted all the levels and lobby files (atleast as far as i remember), although The Line did test an alternative Crab Cat route, so thanks to them for that, even if it didn't end up being used, i really appreciate the help for stuff like that

SMALL NOTE FOR LOBBY FILES: included in the ACAC lobby files is the connection for 17-16 (Dropzle to Coresaken City). This is an normal unused file, but the file in normal unused folder desyncs, so until my resync fix gets added to the main rep, Im adding it to the files here, once its added to main rep ill remove the file from this rep and fix the lobby file, but i didnt want to hold off this draft just for that