Expert Lobby All Cats All Challenges Draft (57:03.902)
In terms of levels, the ones that are different from any% are:
Clockwork
Ethereal Ascension
Meaningless Contraptions
Hypnagogia
Vinculum
A Change in Direction
Overgrown Linn
Caper Cavortion
Hydroshock
Plasma Reactor
Summit Down-Side
Psychokinetic
Garden of Khu'tara
Mosaic Garden
Starfruit Supernova

In terms of the lobby, oh i have a lot to say:
1. As usual, the gym is played first, and THEN the first any% level (Madeline: The Bubble)
2. After M:tB, the TAS goes to the very top-left of the map off of MTB's edge, where the "Madeline: The Bubble" cat is pet.
3. Then, there is the Adventure Cat. Just by looking at the level it starts from and the level it enters, on paper it's not too different, as the routing for Adventure cat uses any% lobby connections so no actual rerouting is needed for it. But how? Well, the entire point of the cat is to adventure around the lobby, and collect 6 door coins, which on the very left edge of the map, there is a little alcove that opens up with a cat. Hence the name, "Adventure Cat." So, after all 6 coins are collected, then what? Well, there happens to be a teleporter right next to the cat. Well, where does it go? Here is where it gets *tricky*:
So the trigger at the teleporter, in essence, wait a certain amount of time after Madeline enters it, then 5 teleport triggers are placed on Madeline at the same time. And, somehow, an RNG function of some sort determines which of these is the one that is used and that randomly determines where she is teleported. One of them teleports Madeline right outside Ethereal Ascension, so we want the RNG to be in our favour, but RNG doesn't necessarily work like that. However, unlike a lot of other random elements in maps, this one is semi-deterministic. As long as it is the first TAS you play after the game starts, this lobby and it's RNG wait will work at that connection. Otherwise, it is random if it works.

As far as I remember, I drafted all of the levels (not incl gym, i drafted none of the gyms)
The Line drafted the two Cat lobby routing files, and figured out how Adventure Cat rng works in terms of the TAS. Thank you to them for the help!

4/5 lobbies, one more to go