Strawberry Jam All Cats All Challenges Draft (3:35:39.380)
I can't believe that this category is technically done! I know sync check will NOT like this file, and for that reason, I would appreciate if anyone had any idea how to: A. fix sync check for beginner-expert gyms (which should fix beginner-expert lobbies (for the most part)), and B. set whatever rng value that makes the Expert Lobby Adventure Cat teleporter randomize its location to its default value upon first booting up the game (as long as youre in the menu, the rng is deterministic, it only changes in the actual gameplay for some reason). 

VERY IMPORTANT TO NOTE: THIS ISN'T ACTUALLY ALL CATS ALL CHALLENGES TECHNICALLY. As unfortunate as it is to say, and trust me it's REALLY unfortunate, this TAS misses a cat and an entire set of challenges: Break My Ivory Tower, AKA BMIT. To make a very annoying story short, its a set of puzzle challenges that you can access either at the end of the level right after the heart, or at the beginning via a secret hidden passage. After completing all 5 BMIT rooms, you get rewarded with a cat, and petting the cat unlocks the Ivory heart in BMIT. This doesn't sound too bad, BUT...
BMIT isn't actually, well, possible technically. By that I mean, there is a solution, and the creator of the map and other people closer to her know the solution and the gimmicks needed, but for most players (incl. TASers), the solution isn't known, only the end goal. I can't TAS something with no known solution whatsoever, so I elected to finish this TAS without BMIT, and add it in LATER once a solution is shared or is found, and I then finish the BMIT TAS.

I can't thank every single individual who helped work on anything ACAC, I am very grateful that my silly little meme TAS idea had interest behind it, it means alot :3

Now, besides BMIT (can't do), all thats left before I take a break from semi-normal SJ TASes (so not including mindash) is max% extended. Most of the work for max-ext is just stitching ACAC with either an ARB or a silver run depending on which is faster, and there isn't a whole lot of like deep routing that comes with it TBH, so it shouldn't take that long to finish realistically, maybe like another week or two depending on how motivated I am.