-2444f Expert Lobby ARB (1:01:31.006->1:00:49.458)
resync, update time, make sure it uses the optimal lobby route (same as any%)
i know what caused the desync for this one, thankfully. the previous evallua command to set debug mode to false didnt work anymore. i replaced it with a set command that seems to work. whether or not it *does* for the bot is another story, but it works for me so...

im assuming itd be a "enforce maingame" thing, if not thatd be annoying and idk how id fix it for full game

also added level names to each level section yippeeeeeeeeee