-2695f Beginner Lobby Max% (1:04:57.896->1:04:12.081)
various improvements to levels, but also a slight reroute where most of it is the same:
it starts on forest path, and continues in the same order after that. then, after the restart chapter after midnight spire, goes to loopy lagoon as the last level, then gets the rainbow by using a snap moonboost that any% uses. gets the rainbow while at decently high speed, and then goes to the same bench as before. continues to heartside asa usual. looks pretty cool tbh! probably improvable

4/5 lobbies done!